requirements.txt
openai>=1.12.0
nvidia-riva-client>=2.14.0
grpcio>=1.60.0
protobuf>=4.25.0
opencv-python>=4.9.0
sounddevice>=0.4.6
numpy>=1.26.0
pyobjc-framework-MapKit>=10.1
pyobjc-framework-CoreLocation>=10.1
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
